Use both TCP and UDP for DNS in firewall example
DNS uses both UDP and TCP. While UDP is usually the default TCP is
needed to handle larger responses. Especially so when the EDNS0
extension isn't available.

Yes, I do realize that this is just an example. Yet, a lot of people
learn from examples, and the misconception that DNS only needs UDP is
already way too common.