Improve performance of images data resource
Previously we were appending all found images to an slice and then
later iterating over that slice to find the right image. This commit
iterates over the results and checks if the image is found in the
current page instead, breaking the loop if the image is found.

Also there was no pagination for the results, so if the image was in
the other than the first page of results it won't be found. This also
addresses and closes #69.