Set resize_disk to the default value on import.
This implements a new `resourceDigitalOceanDropletImport` function
so that we can handle some thing differently while importing
than when using `resourceDigitalOceanDropletRead`.

In particular, this is needed here as `resize_disk` defaults to
`true`, but importing results in an empty value as it a non-API
attribute. So when running `terraform apply` changes are found
regardless of the value in the tf file.

See: https://github.com/terraform-providers/terraform-provider-digitalocean/issues/76