Allows assigning an already existing floating ip to a droplet via a new resource (fixes #8).
Adds a `digitalocean_floating_ip_assignment` resource so that a floating IP that
already exists can be assigned to a droplet.

I've also added the `ignore_droplet_id` flag to the `digitalocean_floating_ip` so that
the existing floating IP resource can be used to create these IPs.

A new resource was used to prevent circular dependencies.