Ensure image ID set to state when importing a Droplet (Fixes: #144).
When importing a Droplet based on an image that no longer
has a slug (or never had one, e.g. a snapshot), the image
ID is not properly being set to state. This had not been
an issue in the past as we did not validate that the image
value was not an empty string.

This PR ensures the correct value is set.