Adding db replica datasource #489 (#608)
* Adding db replica datasource #489

running into a strange "id" attribute issue when testing it

* Trying to fix issue 'no "id" found in attributes' error
error:
=== CONT  TestAccDataSourceDigitalOceanDatabaseReplica_Basic
    testing_new_config.go:67: no "id" found in attributes
    testing_new.go:57: no "id" found in attributes
--- FAIL: TestAccDataSourceDigitalOceanDatabaseReplica_Basic (511.53s)
FAIL

it seems to error on "step 3". When I try to put a breaking point on line
111 of "datasource_digitalocean_database_replica.go" it doesn't seem to reach it.

* fixing race condition of replicas

The database reports as online, but it was too early to use. Used work around by adding sleep timer.

* all TestCheckResourceAttr in step 3 of db rep test pass except "size"

I'd take it's because size isn't set in "resourceDigitalOceanDatabaseReplicaRead" should it be?
"Size" isn't described in the Godo DatabaseReplica struct. I feel like the size of the
replica would be important though to maintain?

* edit attributes of database replica data source

- cut out size
- changed computed to true for optional attributes

* changing resource.ComposeTestCheckFunc to resource.TestCheckFunc

* adding documentation for data source

I only put cluster_id and name as required, and kept tags as optional. I figured the rest (as compared to the resource database replica doc) would be computed?

I'm not 100% sure about the import either.

* fix documentation

Co-authored-by: Andrew Starr-Bochicchio <andrewsomething@users.noreply.github.com>