Change k8s example to use ingress v1
In Kubernetes 1.22, the v1beta1 version of Ingress
is no longer available, thus the example fails to create
the ingress on Kubernetes >= 1.22. Switches the example
to use the kubernetes_ingress_v1 resource instead.

Note that this now increases the minimum version of
terraform-provider-kubernetes since kubernetes_ingress_v1 was
added in v2.7.0.