Rework deployment logic
This commit modifies the logic used by the `digitalocean_app` resource to check each step for a success
status rather than using the counts of steps. The step counting method fails to work with jobs of kind
`FAILED_DEPLOY` as these will always be pending, meaning that any app deployed with a job of this type
will always time out even when it is successful.

Closes #1047

Signed-off-by: David Bond <davidsbond93@gmail.com>